Fix handling of whitespace-only strings and stop stripping empty elements.
Leaf strings that were pure whitespace were not getting removed.
This was because we were setting element.string = "".
Setting element.string does not change the string output when the element is serialised.
We need to instead create a new type(element) object, passing "" to the constructor.

Also removed stripping of empty elements as we were getting odd behaviour.
This behaviour may be due the "next" property of "previous" elements not being updated correctly on extract() and / decompose()